image_watermarking/requirements.txt
TKinter
Pillow
ttf-mscorefonts-installer